requirements.txt
streamlit>=1.32.0
openai>=1.35.0
python-dotenv>=1.0.1
azure-mgmt-web
azure-mgmt-storage
# Azure Cognitive Search SDK (beta) + dependencies
azure-search-documents==11.6.0b12
azure-identity>=1.15.0
azure-core>=1.30.0
azure-ai-agents==1.0.1
azure-ai-projects
azure-ai-documentintelligence

# PDF parsing for the uploader feature in the Streamlit app
pymupdf>=1.24.0

# Additional runtime / notebook helpers required by the MS Quickstart
aiohttp==3.9.0b0
requests>=2.32.0
ipykernel>=6.29.0
tiktoken
azure-keyvault-secrets
langchain_text_splitters
openpyxl
tenacity==8.5.0

# Web and image processing
webvtt-py==0.5.1
Pillow==11.1.0

# Data handling and validation dependencies
jsonschema
tabulate==0.9.0

# HTTP client
httpx>=0.27.0
pandas
PyMuPDF 